+++ cogstat/cogstat_gui.py
@@ -793,7 +793,8 @@ class StatMainWindow(QtWidgets.QMainWindow):
         self._print_to_output_pane()
         self._busy_signal(False)
 
-    def diffusion(self, error_name=[], RT_name=[], participant_name=[], condition_names=[]):
+    def diffusion(self, error_name='', RT_name='', participant_name='', condition_names=[], correct_coding='0',
+                  reaction_time_in='sec'):
         """Run a diffusion analysis on behavioral data.
 
         Arguments:
@@ -808,7 +809,8 @@ class StatMainWindow(QtWidgets.QMainWindow):
             else:
                 self.dial_diffusion.init_vars(names=self.active_data.data_frame.columns)
             if self.dial_diffusion.exec_():
-                error_name, RT_name, participant_name, condition_names = self.dial_diffusion.read_parameters()
+                error_name, RT_name, participant_name, condition_names, correct_coding, reaction_time_in = \
+                    self.dial_diffusion.read_parameters()
             else:
                 return
         self._busy_signal(True)
@@ -819,7 +821,10 @@ class StatMainWindow(QtWidgets.QMainWindow):
                 _('At least the reaction time and the error variables should be given.')))
         else:
             try:
-                text_result = self.active_data.diffusion(error_name, RT_name, participant_name, condition_names)
+                # use the original term in the function call, not the translated one
+                reaction_time_in = 'sec' if reaction_time_in == _('sec') else 'msec'
+                text_result = self.active_data.diffusion(error_name, RT_name, participant_name,
+                                                         condition_names, correct_coding[0], reaction_time_in)
             except:
                 text_result = cs_util.reformat_output(broken_analysis % _('Behavioral data diffusion analysis'))
                 traceback.print_exc()